sleep_dailies_compositor code clean up
